requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
requests
PyMuPDF
GitPython
python-dotenv
